+++ src/main/java/br/com/estoque/dto/auth/RegisterRequestDTO.java
@@ -2,18 +2,21 @@ package br.com.estoque.dto.auth;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
+import io.swagger.v3.oas.annotations.media.Schema;
 
-
-
+@Schema(description = "Dados necessários para registrar um novo usuário")
 public class RegisterRequestDTO {
 	
+	@Schema(description = "Nome do usuário", example = "Davi Esteves")
 	@NotBlank(message = "Nome obrigatório")
 	private String nome; 
 	
+	@Schema(description = "Email usado para login", example = "davi@email.com")
 	@Email(message = "Email deve ser válido")
 	@NotBlank(message = "Email é obrigatório")
 	private String email;
 	
+    @Schema(description = "Senha do usuário com no mínimo 8 caracteres", example = "12345678")
 	@Size(min = 8, message = "Senha precisa ter no mínimo 8 caractéres")
 	@NotBlank(message = "Senha é obrigatória")
 	private String senha;